+++ jukebox-app/src/main/java/se/qxx/android/jukebox/activities/fragments/JukeboxPlayerFragment.java
@@ -1,0 +1,109 @@
+package se.qxx.android.jukebox.activities.fragments;
+
+
+import android.os.Bundle;
+import android.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import com.google.protobuf.RpcCallback;
+
+import se.qxx.android.jukebox.R;
+import se.qxx.jukebox.domain.JukeboxDomain;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class JukeboxPlayerFragment extends RemotePlayerFragment {
+
+
+    public JukeboxPlayerFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        TextView textView = new TextView(getActivity());
+        textView.setText(R.string.hello_blank_fragment);
+        return textView;
+    }
+
+    @Override
+    public RpcCallback<JukeboxDomain.JukeboxResponseStartMovie> getCallback() {
+        return null;
+    }
+
+    @Override
+    public void setSubtitle(JukeboxDomain.SubtitleUri subtitleUri) {
+
+    }
+
+    @Override
+    public void start() {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void stop() {
+
+    }
+
+    @Override
+    public int getDuration() {
+        return 0;
+    }
+
+    @Override
+    public int getCurrentPosition() {
+        return 0;
+    }
+
+    @Override
+    public void seekTo(int pos) {
+
+    }
+
+    @Override
+    public boolean isPlaying() {
+        return false;
+    }
+
+    @Override
+    public int getBufferPercentage() {
+        return 0;
+    }
+
+    @Override
+    public boolean canPause() {
+        return false;
+    }
+
+    @Override
+    public boolean canSeekBackward() {
+        return false;
+    }
+
+    @Override
+    public boolean canSeekForward() {
+        return false;
+    }
+
+    @Override
+    public boolean isFullScreen() {
+        return false;
+    }
+
+    @Override
+    public void toggleFullScreen() {
+
+    }
+}
